fix: use streaming methods where possible
The base adapter class calls `.put` repeatedly when `.putMany` is invoked
in order to reduce the number of methods a datastore has to implement to
be compatible with the datastore interface.

This hurts performance where datastores wrap other datastores that can
take advantage of the streaming methods to exectute certain operations in
parallel.

The change here is to call streaming methods from streaming methods where
it makes sense to do so, instead of letting the adapter class convert
streaming method invocations into serial method invocations.